Added separate module 'rich_amami.py' to set aparameters for rich and rich-argparse
